feat: adding custom responses

+++ api/src/models/User.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID        uint   `json:id,omitempty`
+	ID        uint64 `json:id,omitempty`
 	Nome      string `json:nome,omitempty`
 	Nick      string `json:nick,omitempty`
 	Email     string `json:email,omitempty`
